+++ tests/testthat/tests.R
@@ -8,7 +8,7 @@ test_that("amce() works", {
 })
 
 test_that("amce_diffs() works", {
-    expect_true(inherits(x <- amce_diffs(hainmueller, ChosenImmigrant ~ Gender, id = ~ CaseID), "cj_mm"))
+    expect_true(inherits(x <- amce_diffs(hainmueller, ChosenImmigrant ~ LanguageSkills, by = ~ Gender, id = ~ CaseID), "cj_diffs"))
     expect_true(inherits(plot(x), "ggplot"))
 })
 
@@ -29,7 +29,6 @@ test_that("freqs() works", {
 
 test_that("props() works", {
     expect_true(inherits(x <- props(hainmueller, ChosenImmigrant ~ Gender, id = ~ CaseID), "cj_props"))
-    expect_true(inherits(plot(x), "ggplot"))
 })
 
 test_that("cj_anova() works", {