+++ R/mm.R
@@ -21,7 +21,7 @@
 #' data(hainmueller)
 #' mm(hainmueller, ChosenImmigrant ~ Gender + Education + LanguageSkills,
 #'    id = ~ CaseID, h0 = 0.5)
-#' @seealso \code{\link{plot.cj_mm}}
+#' @seealso \code{\link{mm_diffs}} \code{\link{plot.cj_mm}}
 #' @import stats
 #' @importFrom survey svydesign svyby svymean
 #' @export