+++ NEWS.md
@@ -1,6 +1,8 @@
 # cregg 0.1.8
 
 * `mm()` gains an `h0` argument to specify a null hypothesis values so that z statistics and p-values are meaningful.
+* Cleaned up documentation and expanded 'Introduction' vignette, moving README content to there.
+* Require **survey** version 0.33 (for `family` argument).
 * Added a basic test suite. (#4)
 
 # cregg 0.1.7